Implement operation contract system for dtype and ndim inference
- Introduced  and  enums to define how operations affect data types and dimensions.
- Created  class to declare the effects of operations on dtype and ndim, ensuring consistency with Rust's execution layer.
- Updated the  class to utilize  for static type inference, handling special cases for parameter-dependent overrides.
- Enhanced dtype validation in Rust to ensure output matches the planned dtype from Python contracts.
- Added comprehensive tests to verify the correctness of the operation contract system and dtype preservation across various operations.

+++ view-buffer/src/expr.rs
@@ -499,7 +499,7 @@ impl ViewExpr {
         })
     }
 
-    pub fn threshold(self: &Arc<Self>, value: u8) -> Arc<Self> {
+    pub fn threshold(self: &Arc<Self>, value: f64) -> Arc<Self> {
         let op = ImageOp {
             kind: ImageOpKind::Threshold(value),
         };
